requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
python-dotenv==1.0.0
sqlalchemy==2.0.20
alembic==1.12.0
pydantic==2.3.0
pydantic-settings==2.0.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
httpx==0.24.1
numpy==1.25.2
pandas==2.0.3
scikit-learn==1.3.0
scipy==1.11.2
python-dateutil==2.8.2
pytest==7.4.0
pytest-cov==4.1.0
